+++ app/mission/imap/+mag/+app/+imap/+control/CPT.m
@@ -37,7 +37,7 @@ classdef CPT < mag.app.Control & mag.app.mixin.Filter
             primaryModeLabel.Layout.Column = 1;
 
             this.PrimaryModePatternField = uieditfield(this.Layout, Value = this.encodeForEditField(this.PrimaryModePattern), ...
-                ValueChangingFcn = @(~, value) this.validatePattern(value));
+                ValueChangedFcn = @(~, value) this.validatePattern(value));
             this.PrimaryModePatternField.Layout.Row = 2;
             this.PrimaryModePatternField.Layout.Column = [2, 3];
 
@@ -48,7 +48,7 @@ classdef CPT < mag.app.Control & mag.app.mixin.Filter
             secondryModeLabel.Layout.Column = 1;
 
             this.SecondaryModePatternField = uieditfield(this.Layout, Value = this.encodeForEditField(this.SecondaryModePattern), ...
-                ValueChangingFcn = @(~, value) this.validatePattern(value));
+                ValueChangedFcn = @(~, value) this.validatePattern(value));
             this.SecondaryModePatternField.Layout.Row = 3;
             this.SecondaryModePatternField.Layout.Column = [2, 3];
 
@@ -59,7 +59,7 @@ classdef CPT < mag.app.Control & mag.app.mixin.Filter
             rangeLabel.Layout.Column = 1;
 
             this.RangePatternField = uieditfield(this.Layout, Value = this.encodeForEditField(this.RangePattern), ...
-                ValueChangingFcn = @(~, value) this.validatePattern(value));
+                ValueChangedFcn = @(~, value) this.validatePattern(value));
             this.RangePatternField.Layout.Row = 4;
             this.RangePatternField.Layout.Column = [2, 3];
         end
@@ -99,7 +99,7 @@ classdef CPT < mag.app.Control & mag.app.mixin.Filter
             pattern = asManyOfPattern(digitsPattern() + optionalPattern(characterListPattern(",") + whitespacePattern()));
 
             if ~matches(value, pattern)
-                error("Value must match the pattern '1, 2, 3'.");
+                error("mag:app:invalidPattern", "Value must match the pattern ""1, 2, 3"".");
             end
         end
     end